Update to store machine learning models in pickled format and just load them to perform predictions, rather than re-training them.
