+++ backend/app.yaml
@@ -1,3 +1,4 @@
 runtime: python39
 service: api
-instance_class: F4_1G
+instance_class: F2
+
