+++ backend/main.py
@@ -2,11 +2,9 @@ from flask import Flask, request, jsonify, render_template
 from players import get_all_players
 from predictions import win_prediction, total_points_prediction, number_of_games_prediction
 from predictions import games_decided_by_extra_points_prediction, first_game_winner_prediction
-from predictions import money_line_prediction
-from machine_learning.tennis import fetch_tournament
+from old_scraper.tennis import fetch_tournament
 from users import get_all_users, add_user, user_exists
 from player_statistics import generate_player_statistics, get_player_stats, get_vs_stats
-from predictions_final import initialize_predictions
 from flask_cors import CORS
 
 app = Flask(__name__)
